// pop-web/src/app/shared/openapi/model/rai-staging.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { CodedConcept } from './coded-concept';


export interface RaiStaging { 
    stagingDomain?: RaiStaging.StagingDomainEnum;
    /**
     * 
     */
    id: string;
    /**
     * 
     */
    createdAt: string;
    /**
     * 
     */
    updatedAt: string;
    createdBy?: User | null;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s cancer is staged
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the staging was performed and recorded.
     */
    date: string;
    /**
     * The user(s) who updated the data since its creation
     */
    updatedBy?: Array<User>;
    /**
     * References to the neoplastic entities that were the focus of the staging.
     */
    stagedEntitiesIds?: Array<string>;
    /**
     * The value of the Rai stage
     */
    stage: CodedConcept;
    methodology?: CodedConcept | null;
    /**
     * Human-readable description
     */
    description: string;
}
export namespace RaiStaging {
    export type StagingDomainEnum = 'rai';
    export const StagingDomainEnum = {
        Rai: 'rai' as StagingDomainEnum
    };
}